% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx-bbox.R
\name{crs_approx_bbox}
\alias{crs_approx_bbox}
\title{Approximate bounding box transform}
\usage{
crs_approx_bbox(
  handleable,
  crs_to,
  crs_from = wk::wk_crs(handleable),
  engine = crs_default_engine(),
  n = 20,
  ...
)
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[wk:wkb]{wkb()}}, \code{\link[wk:wkt]{wkt()}}, \code{\link[wk:xy]{xy()}},
\code{\link[wk:rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[wk:wk_handle]{wk_handle()}} is defined.}

\item{crs_to}{Source and destination coordinate reference systems}

\item{crs_from}{Source and destination coordinate reference systems}

\item{engine}{A transform engine such as \code{\link[=crs_engine_null]{crs_engine_null()}}}

\item{n}{The number of points in a grid used to approximate the transform}

\item{...}{engine-specific transformation options}
}
\value{
A \code{\link[wk:rct]{wk::rct()}} representing the transformed bounding box
}
\description{
Approximate bounding box transform
}
\examples{
crs_approx_bbox(
  wk::rct(-180, -1, 180, 1, crs = "OGC:CRS84"), "OGC:CRS84",
  engine = crs_engine_identity()
)

}
