% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-engine.R
\name{crs_engine}
\alias{crs_engine}
\alias{crs_set_engine}
\alias{with_crs_engine}
\alias{is_crs_engine}
\title{Default coordinate transform engine}
\usage{
crs_engine(env = parent.frame())

crs_set_engine(engine, env = parent.frame())

with_crs_engine(engine, expr, env = parent.frame())

is_crs_engine(engine, env = parent.frame())
}
\arguments{
\item{env}{The environment to look for named functions if the engine
option is a string.}

\item{engine}{A transform engine such as \code{\link[=crs_engine_null]{crs_engine_null()}}}

\item{expr}{The expression to evaluate with the transform engine}
}
\value{
\itemize{
\item \code{crs_engine()} returns the current default transform engine
\item \code{crs_set_engine()} returns the previously set default transform
engine
\item \code{with_crs_engine()} returns the result of \code{expr}
}
}
\description{
Default coordinate transform engine
}
\examples{
crs_engine()
prev_engine <- crs_set_engine(crs_engine_identity())
# ...
crs_set_engine(prev_engine)

with_crs_engine(crs_engine_identity(), {
  # ...
})

}
