% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine-fun.R
\name{crs_engine_fun}
\alias{crs_engine_fun}
\alias{crs_engine_fun_define}
\alias{crs_engine_fun_get}
\alias{crs_engine_get_wk_trans.crs2crs_engine_fun}
\title{Transform using R functions}
\usage{
crs_engine_fun()

crs_engine_fun_define(engine, crs_to, crs_from, fun)

crs_engine_fun_get(engine, crs_to, crs_from)

\method{crs_engine_get_wk_trans}{crs2crs_engine_fun}(engine, handleable, crs_to, crs_from)
}
\arguments{
\item{engine}{A \code{\link[=crs_engine_fun]{crs_engine_fun()}}}

\item{crs_to}{Source and destination coordinate reference systems}

\item{crs_from}{Source and destination coordinate reference systems}

\item{fun}{A function that accepts the output of \code{\link[wk:wk_vertices]{wk::wk_coords()}} and modifies
the \code{x}, \code{y}, \code{z}, and/or \code{m} columns. These columns can also be added
or removed to set or drop the dimensions of the output.}

\item{handleable}{A geometry vector (e.g., \code{\link[wk:wkb]{wkb()}}, \code{\link[wk:wkt]{wkt()}}, \code{\link[wk:xy]{xy()}},
\code{\link[wk:rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[wk:wk_handle]{wk_handle()}} is defined.}
}
\value{
\itemize{
\item \code{crs_engine_fun()}: An engine that can be used for transforms
\item \code{crs_engine_fun_define()}: Returns \code{engine}, modified mutably
\item \code{crs_engine_fun_get()}: A \code{\link[=function]{function()}} that accepts a single argument
that is the output of \code{\link[wk:wk_vertices]{wk::wk_coords()}}.
}
}
\description{
Transform using R functions
}
\examples{
engine <- crs_engine_fun()
engine <- crs_engine_fun_define(engine, "EPSG:3857", "OGC:CRS84", function(coords) {
  r <- 6378137
  coords$x <- coords$x * pi / 180 * r
  coords$y <- log(tan(pi / 4 + coords$y * pi / 180 / 2)) * r
  coords
})

crs_engine_transform(engine, wk::xy(-64, 45, crs = "OGC:CRS84"), "EPSG:3857")

}
