% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine-proj-sf.R
\name{crs_engine_proj_sf}
\alias{crs_engine_proj_sf}
\alias{crs_engine_proj_pipeline.crs2crs_engine_proj_sf}
\alias{crs_engine_proj_pipeline_apply.crs2crs_engine_proj_sf}
\alias{crs_engine_transform.crs2crs_engine_proj_sf}
\title{Use PROJ via sf to transform coordinates}
\usage{
crs_engine_proj_sf()

\method{crs_engine_proj_pipeline}{crs2crs_engine_proj_sf}(
  engine,
  handleable,
  crs_to,
  crs_from = wk::wk_crs(handleable),
  bbox = wk::wk_bbox(handleable),
  ...
)

\method{crs_engine_proj_pipeline_apply}{crs2crs_engine_proj_sf}(engine, handleable, pipeline, ...)

\method{crs_engine_transform}{crs2crs_engine_proj_sf}(
  engine,
  handleable,
  crs_to,
  crs_from = wk::wk_crs(handleable),
  ...
)
}
\arguments{
\item{engine}{A transform engine such as \code{\link[=crs_engine_null]{crs_engine_null()}}}

\item{handleable}{A geometry vector (e.g., \code{\link[wk:wkb]{wkb()}}, \code{\link[wk:wkt]{wkt()}}, \code{\link[wk:xy]{xy()}},
\code{\link[wk:rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[wk:wk_handle]{wk_handle()}} is defined.}

\item{crs_to}{Source and destination coordinate reference systems}

\item{crs_from}{Source and destination coordinate reference systems}

\item{bbox}{The optional bounding box of the object.
Defaults to \code{\link[wk:wk_bbox]{wk::wk_bbox()}} of \code{handleable} forced to \code{crs_from}.
Use \code{NULL} to skip bounding box selectio for a single transformation.}

\item{...}{Arguments passed on to \code{\link[sf:st_transform]{sf::st_transform()}}}

\item{pipeline}{A PROJ coordinate transformation pipeline definition}
}
\value{
\itemize{
\item \code{crs_engine_proj_sf()} returns an engine that can be used to transform coordinates
}
}
\description{
The PROJ sf engine is a thin wrapper around \code{\link[sf:st_transform]{sf::st_transform()}}
using \code{\link[sf:proj_tools]{sf::sf_proj_pipelines()}} to query the most appropriate transformation.
The interface is similar to that of \code{\link[=crs_engine_proj_cmd]{crs_engine_proj_cmd()}} to allow a
drop-in replacement for most use-cases.
}
\examples{
if (requireNamespace("sf", quietly = TRUE)) {
  engine <- crs_engine_proj_sf()
  crs_transform(
    wk::xy(-64, 45, crs = "OGC:CRS84"), "EPSG:3857",
    engine = engine
  )
}

}
