% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine-sf.R
\name{crs_engine_sf}
\alias{crs_engine_sf}
\alias{crs_has_sf}
\alias{crs_engine_transform.crs2crs_engine_sf}
\title{Use sf to transform coordinates}
\usage{
crs_engine_sf(authority_compliant = NA)

crs_has_sf()

\method{crs_engine_transform}{crs2crs_engine_sf}(
  engine,
  handleable,
  crs_to,
  crs_from = wk::wk_crs(handleable),
  ...
)
}
\arguments{
\item{authority_compliant}{Use \code{TRUE} or \code{FALSE} to apply \code{\link[sf:st_transform]{sf::st_transform()}}
with \code{\link[sf:st_crs]{sf::st_axis_order()}} temporarily set. The default \code{NA} uses whatever
the current value of \code{\link[sf:st_crs]{sf::st_axis_order()}} happens to be at the time of
the transform.}

\item{engine}{A transform engine such as \code{\link[=crs_engine_null]{crs_engine_null()}}}

\item{handleable}{A geometry vector (e.g., \code{\link[wk:wkb]{wkb()}}, \code{\link[wk:wkt]{wkt()}}, \code{\link[wk:xy]{xy()}},
\code{\link[wk:rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[wk:wk_handle]{wk_handle()}} is defined.}

\item{crs_to}{Source and destination coordinate reference systems}

\item{crs_from}{Source and destination coordinate reference systems}

\item{...}{Arguments passed on to \code{\link[sf:st_transform]{sf::st_transform()}}}
}
\value{
\itemize{
\item \code{crs_engine_sf()} returns an engine that can be used to transform coordinates
}
}
\description{
The sf engine is a thin wrapper around \code{\link[sf:st_transform]{sf::st_transform()}} but preserving
the output type to mach that of the input. Using the sf engine allows you
to use \code{\link[=crs_transform]{crs_transform()}} as a drop-in replacement for \code{\link[sf:st_transform]{sf::st_transform()}}.
The sf engine is not authority compliant (i.e., geodedic coordinates are
assumed to have lon,lat axis order).
}
\examples{
if (crs_has_sf()) {
  engine <- crs_engine_sf()
  crs_transform(
    wk::xy(-64, 45, crs = "OGC:CRS84"), "EPSG:3857",
    engine = engine
  )
}

}
