% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{crs_transform}
\alias{crs_transform}
\alias{crs_transform_pipeline}
\alias{crs_set}
\alias{crs_get}
\title{Transform Coordinate Reference Systems}
\usage{
crs_transform(
  handleable,
  crs_to,
  crs_from = wk::wk_crs(handleable),
  engine = crs_engine(),
  ...
)

crs_transform_pipeline(handleable, pipeline, engine = crs_engine(), ...)

crs_set(handleable, crs_to)

crs_get(handleable)
}
\arguments{
\item{handleable}{A geometry vector (e.g., \code{\link[wk:wkb]{wkb()}}, \code{\link[wk:wkt]{wkt()}}, \code{\link[wk:xy]{xy()}},
\code{\link[wk:rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[wk:wk_handle]{wk_handle()}} is defined.}

\item{crs_to}{Source and destination coordinate reference systems}

\item{crs_from}{Source and destination coordinate reference systems}

\item{engine}{A transform engine such as \code{\link[=crs_engine_null]{crs_engine_null()}}}

\item{...}{engine-specific transformation options}

\item{pipeline}{A character vector representing a PROJ coordinate operation}
}
\value{
A modified \code{handleable} with \code{crs_to}.
}
\description{
Transform Coordinate Reference Systems
}
\examples{
engine <- crs_engine_fun()
engine <- crs_engine_fun_define(engine, "EPSG:3857", "OGC:CRS84", function(coords) {
  r <- 6378137
  coords$x <- coords$x * pi / 180 * r
  coords$y <- log(tan(pi / 4 + coords$y * pi / 180 / 2)) * r
  coords
})

obj <- wk::xy(-64, 45, crs = "OGC:CRS84")
crs_transform(obj, "EPSG:3857", engine = engine)

with_crs_engine(engine, {
  crs_transform(obj, "EPSG:3857")
})

}
