% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs2crs-package.R
\docType{package}
\name{crs2crs-package}
\alias{crs2crs}
\alias{crs2crs-package}
\title{crs2crs: Generic Coordinate System Transforms}
\description{
Provides a flexible framework for reproducible
  coordinate transforms.
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
