% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine-fun.R
\name{crs_engine_fun}
\alias{crs_engine_fun}
\alias{crs_engine_fun_define}
\alias{crs_engine_fun_get}
\alias{crs_transform_fun}
\alias{crs_engine_get_wk_trans.crs2crs_engine_fun}
\title{Transform coordinates using R functions}
\usage{
crs_engine_fun(prepare = wk::wk_crs_proj_definition)

crs_engine_fun_define(engine, crs_to, crs_from, fun)

crs_engine_fun_get(engine, crs_to, crs_from)

crs_transform_fun(handleable, fun)

\method{crs_engine_get_wk_trans}{crs2crs_engine_fun}(engine, handleable, crs_to, crs_from, ...)
}
\arguments{
\item{prepare}{A function that converts CRS objects to a canonical string
definition. This reduces the number of function definitions that might
otherwise be required to capture the transforms. The default is to
use \code{\link[wk:wk_crs_proj_definition]{wk::wk_crs_proj_definition()}}, which returns the shortest possible
definition of a CRS that can be used as a PROJ definition
(usually something like "EPSG:32620"), falling back to a more verbose
definition if an authority string is not found.}

\item{engine}{A \code{\link[=crs_engine_fun]{crs_engine_fun()}}}

\item{crs_to}{Source and destination coordinate reference systems}

\item{crs_from}{Source and destination coordinate reference systems}

\item{fun}{A function that accepts the output of \code{\link[wk:wk_vertices]{wk::wk_coords()}} and modifies
the \code{x}, \code{y}, \code{z}, and/or \code{m} columns. These columns can also be added
or removed to set or drop the dimensions of the output.}

\item{handleable}{A geometry vector (e.g., \code{\link[wk:wkb]{wkb()}}, \code{\link[wk:wkt]{wkt()}}, \code{\link[wk:xy]{xy()}},
\code{\link[wk:rct]{rct()}}, or \code{\link[sf:sfc]{sf::st_sfc()}}) for which \code{\link[wk:wk_handle]{wk_handle()}} is defined.}

\item{...}{engine-specific transformation options}
}
\value{
\itemize{
\item \code{crs_engine_fun()}: An engine that can be used for transforms
\item \code{crs_engine_fun_define()}: Returns \code{engine}, modified mutably
\item \code{crs_engine_fun_get()}: A \code{\link[=function]{function()}} that accepts a single argument
that is the output of \code{\link[wk:wk_vertices]{wk::wk_coords()}}.
\item \code{crs_transform_fun()}: Returns a modified \code{handleable} with \code{fun} applied
to the coordinates
}
}
\description{
Transform coordinates using R functions
}
\examples{
engine <- crs_engine_fun()
engine <- crs_engine_fun_define(engine, "EPSG:3857", "OGC:CRS84", function(coords) {
  r <- 6378137
  coords$x <- coords$x * pi / 180 * r
  coords$y <- log(tan(pi / 4 + coords$y * pi / 180 / 2)) * r
  coords
})

crs_transform(
  wk::xy(-64, 45, crs = "OGC:CRS84"),
  "EPSG:3857",
  engine = engine
)

# can also use a function to apply a generic transform
crs_transform_fun(wk::xy(1, 2), function(coords) {
  coords$x <- coords$x * 2
  coords$y <- coords$y * 2
  coords
})

}
