% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{crs_hash_prepare}
\alias{crs_hash_prepare}
\alias{crs_hash_prepare.default}
\title{CRS object generic methods}
\usage{
crs_hash_prepare(crs)

\method{crs_hash_prepare}{default}(crs)
}
\arguments{
\item{crs}{An arbitrary R object}
}
\value{
\itemize{
\item \code{crs_hash_prepare()} Returns an object that will be passed to
\code{\link[rlang:hash]{rlang::hash()}} to identify a CRS object. The object returned by
this method should be canonical.
}
}
\description{
CRS object generic methods
}
\examples{
rlang::hash(crs_hash_prepare(1234))

}
